Move SavedInput to src/meta/collections
To make it safe to import by src/meta/data's imports
