+++ packages/db/src/meta/collections.ts
@@ -93,6 +93,15 @@ export type Input<
   N extends CollectionName<C> = CollectionName<C>
 > = CollectionProperty<"input", C, N>;
 
+export type SavedInput<
+  C extends Collections = Collections,
+  N extends CollectionName<C> = CollectionName<C>
+> = {
+  [K in keyof Input<C, N> | "id"]: K extends keyof Input<C, N>
+    ? Input<C, N>[K]
+    : string;
+};
+
 export type IdFields<
   C extends Collections = Collections,
   N extends CollectionName<C> = CollectionName<C>
