+++ packages/db/src/meta/data.ts
@@ -6,10 +6,10 @@ import PouchDB from "pouchdb";
 import type {
   Collections,
   CollectionName,
-  Input,
   MutableCollectionName,
   MutationInput,
-  MutationPayload
+  MutationPayload,
+  SavedInput
 } from "./collections";
 import type { IdObject } from "./id";
 
@@ -43,15 +43,6 @@ export interface Workspace<C extends Collections> {
     input: MutationInput<C, M>
   ): Promise<void>;
 }
-
-export type SavedInput<
-  C extends Collections = Collections,
-  N extends CollectionName<C> = CollectionName<C>
-> = {
-  [K in keyof Input<C, N> | "id"]: K extends keyof Input<C, N>
-    ? Input<C, N>[K]
-    : string;
-};
 
 export type History = PouchDB.Core.IdMeta & PouchDB.Core.GetMeta;
