+++ packages/db/src/resources/index.ts
@@ -49,6 +49,7 @@ export {
   MutableCollectionName,
   NamedCollectionName,
   Input,
+  SavedInput,
   Resource,
   IdFields,
   IdObject,
